requirements.txt
-e .
pytest
aws-cdk.aws-cloudformation
aws-cdk.aws-codepipeline
aws-cdk.aws-codepipeline-actions
aws_cdk.aws_codebuild
aws_cdk.aws_ecs_patterns
aws_cdk.aws_iam